+++ tests/Equinox.CosmosStore.Integration/AccessStrategies.fs
@@ -32,7 +32,7 @@ module WiringHelpers =
 /// This is especially relevant when events are spread between a Tip page and preceding pages as the Tip reading logic is special cased compared to querying
 module SequenceCheck =
 
-    let streamId = Equinox.StreamId.map "_SequenceCheck" (fun (g : Guid) -> g.ToString "N")
+    let target = Equinox.Target.gen "_SequenceCheck" (fun (g : Guid) -> g.ToString "N")
 
     module Events =
 
@@ -69,7 +69,7 @@ module SequenceCheck =
             decider.Transact(decide (value, count), id)
 
     let private create resolve =
-        Service(streamId >> resolve)
+        Service(target >> resolve)
 
     module Config =