+++ tests/Equinox.EventStore.Integration/StoreIntegration.fs
@@ -508,7 +508,7 @@ type Tests(testOutputHelper) =
         let context = createContext connection batchSize
         let id = Guid.NewGuid()
         let toStreamId (x : Guid) = x.ToString "N"
-        let decider = SimplestThing.resolve log context (Equinox.StreamId.map SimplestThing.Category toStreamId id)
+        let decider = SimplestThing.resolve log context (Equinox.Target.gen SimplestThing.Category toStreamId id)
 
         let! before, after = decider.TransactEx(
             (fun state -> state.Version, [SimplestThing.StuffHappened]),